run: update the special fields regexp to match also systrace embedded traces
The catapult/systrace.py tool collects from the target a textual trace
which is embedded in the generated TraceView HTML format report file.
These events have a slightly different format with respect to the ones
normally generated by "trace-cmd report".

This patch updates the regexp used to match special files of each event
to match both the trace-cmd as well as the systrace used format.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>