run: add support to parse systrace.py embedded traces
The systrace.py tool embeds textual traces which are collected by reading
directly the /sys/kernel/debug/tracing/trace file.

In these traces, at least on Android targets based on kernel v3.18,
sched_switch events are generated using a format string which is similar
to the raw format (i.e. sequence of key=value pairs) but includes a
single "==>" substring between the exiting and entering task.

This patch adds the required support to identify when we are parsing
a systrace file, which is expected to have the ".html" extension.
Furthermore, it adds a specific regexp into the Run object parsing code
to remove from sched_switch events the not necessary " ==>" string.
Once this replacement has been done, the event can be properly
parsed as a standard raw event.

A more generic approach which, for example, allows each registered/enabled
class to pre-process the event input line before passing it to the tokenizer
could be probably a better approach. However, at least for pre-defined events,
this could eventually lead to higher overheads.
Such a support could be used to support a further additional feature which
auto-detect the format of the input trace and properly re-format the
data string into the canonical "key=value" form.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>