run: identify CPUs count from dataset if not reported by trace medatada
Some traces, e.g. a systrace.py embedded trace, do not report the same
medatata generated by trace-cmd report. The "_cpus" attribute is required
by plot_trace to properly dimension the plot.

This patch ensure this attribute is available by eventually estimating the
number of CPUs from the events available in the trace.
The sched_switch event is always enable, thus it should be good enough
to provide a proper initialization in all the cases where the CPUs number
is not reported in the header. So fare these cases are just systrace.py
embedded traces.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>