run: use itertools to filter the interesting lines in the trace
Filtering the lines that contain the unique word is the part of the code
where we spend most of the time when parsing the trace.  By doing it in
python we add too much interpreter overhead.  Offload the loop to
itertools (which is in C) to improve performance.

On my computer, parsing a 27Mb trace.dat is 2.23x faster.