Refactred to group series (constraints) on each axis
This is usefull for plots where you need to know the number of series on an axis
beforehand e.g. a bar plot.

Also, the title is now set in the subclass (LinePlot) allowing a specific plot to
easily modify the titles if needed.

Old plot method has been left until _resolve_concat is modified to use plot_axis.