plotter: Properly catch multiple exceptions in Constraint.py
except FooError, BarError:

Doesn't catch both FooError and BarError.  That code catches FooError
and assigns the catched exception to a variable called BarError.  This
is fixed in Python 3 because of that.  The correct syntax to catch
multiple exceptions is to pass the as the first argument to the except
keyword as a tuple.