ILinePlot: put javascript after the table tag
Currently ILinePlot embeds the requirejs call that generates the dygraph
graph in the div that will contain the graph.  Dygraph puts its
generated html in that div and in the legend div that is present after
this one.  This is not a problem for jupyter notebooks because
by the time requirejs gets to run its function and calls
ILinePlot.generate() the DOM has been fully populated so dygraph can
replace the contents of the fig and legend divs without problems.

However, for exported notebooks this becomes a problem.  As
ILinePlot.generate() is now run when parsing the HTML, dygraph gets
called before the legend div is generated and it fails to find it in the
DOM.  Let the HTML create the table with empty divs and move all the
javascript code after the table for dygraph to find the DOM nodes that
it needs for exported notebooks.