trappy/parsing: Ensure parse_raw is set for sched_switch events
When the sched_switch event is added to the parser the constructor is
not called so we never set the parse_raw flag. Ensure there is a class
variable to hold the correct value and use it in the constructor.

We use the raw file to extract the events since we actually do construct
one of these objects during parsing.

Signed-off-by: Chris Redpath <chris.redpath@arm.com>