trappy: Add support for fall-back behaviour for 'print:' event
This is just like tracing_mark_write but with a different name. It
poses a minor difficulty as its name is a Python 2 keyword (so
'ftrace.print.data_frame' is a syntax error) - this is solved by
appending an underscore to the name, as suggested by PEP8.

A base class is added which encapsulates the fallback behaviour for
both tracing_mark_write and print.