trappy/cache: Store the csv in utf-8 charset explicitly
When we load the files from trace-cmd output, we explicitly choose
to load them using utf-8. This means we may have characters in
our trace which cannot be represented in ascii, which makes pandas
upset when we try to store the parsed dataframes in the csv cache.

This is not an issue for Python3 versions, as the to_csv method
defaults to utf-8, whilst the Python2 to_csv method defaults to
ascii.

Signed-off-by: Chris Redpath <chris.redpath@arm.com>