tests/test_ftrace: Fix test_get_all_freqs_data()
In Python2, iterating over a dict's keys will always give the same
sequence (although the ordering is arbitrary). In Python3 that
ordering is still abitrary, but to make things even better it can
change from one execution to another - see [1].

GenericFTrace.get_all_freqs_data() returns a list which's ordering
depends on the keys of the input dictionnary. That in itself would
need to be changed to a better design, but for the time being we can
fix the test and make Trappy pass reliably by converting the output
of get_all_freqs_data() to a dict().

[1]: https://docs.python.org/2/using/cmdline.html#cmdoption-r