base: Tune down string to int conversion
Conversion of hexadecimal-containg strings was fixed in

  commit 9eca88cf9dac ("base: Fix string to int conversion")

but was made slightly overzealous. An event field that really is a
string but that is a valid hexadecimal value without leading "0x" is
currently converted to an int.

Turns out that passing base=0 to int() means to let python figure out
the base from the shape of the string, so let's do that instead of
trying to be too smart for our own good.