base.py: handle any duplicated indices when creating dataframe
When running Lisa tests on fastmodels duplicated timestamps happen
frequently and cause some failures down the line.

See https://github.com/ARM-software/lisa/pull/864 for more details.

The duplicated timestamps are not bad themselves and a side effect of
simulated time not having good enough resolution.

Handling this duplication at the dataframe creation seems to be the most
generic and future proof solution.

The change breaks the test_idle.py because it has a duplicate entry in
its test vector and checks for an exact match. Since we know now that
duplicates will be removed when creating the dataframe, we modify the
test to remove the duplicate value.

Signed-off-by: Qais Yousef <qais.yousef@arm.com>