utils.py: take __line into account in handle_duplicate_index()
To ensure that we don't mess up the ordering of the events when fixing
up the duplicate indices, sort the values first by 'Time' (which is the
index) and then '__line' columns.

If '__line' column doesn't exist then we carry on anyways assuming we
can't do much preserving the original behavior of the function.

Signed-off-by: Qais Yousef <qais.yousef@arm.com>