ftrace: Ensure timestamp uniqueness
The timestamp of consecutive events are sometimes the same, if they were emitted
in a very short interval of time. This becomes an issue when the events are
dispatched in separate dataframes, which then sometimes need to be joined back
together in client code. At this point, the ordering between events can be lost,
leading to wrong results when exact ordering is required.

In order to fix that, pick the next representable floating value when there is
an ambiguity. This ensures that the delta added is the smallest possible, which
is around the 16th least significant digit.